Feat(zone): Implementada função de localizar zona do usuário

+++ lib/features/zone/domain/models/zone.dart
@@ -4,7 +4,7 @@ class Zone{
 	final String name;
 	final double latitude;
 	final double longitude;
-	final double radiusKm;
+	final int radiusKm;
 
 	Zone({
 		required this.id,
